backend/requirements.txt
llama-index
langchain
pypdf2
python-dotenv
uvicorn
fastapi
fastapi-cors